requirements.txt
pyopengl
numpy
glfw
pillow
PyGLM